Refactor AbstractLfBus (#123)
