+++ src/main/java/com/powsybl/openloadflow/network/impl/LfDanglingLineBus.java
@@ -7,15 +7,13 @@
 package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.DanglingLine;
-import com.powsybl.openloadflow.network.AbstractFictitiousLfBus;
-import com.powsybl.openloadflow.network.PerUnit;
 
 import java.util.Objects;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfDanglingLineBus extends AbstractFictitiousLfBus {
+public class LfDanglingLineBus extends AbstractLfBus {
 
     private final DanglingLine danglingLine;
 
@@ -25,6 +23,8 @@ public class LfDanglingLineBus extends AbstractFictitiousLfBus {
         super(Networks.getPropertyV(danglingLine), Networks.getPropertyAngle(danglingLine));
         this.danglingLine = Objects.requireNonNull(danglingLine);
         nominalV = danglingLine.getTerminal().getVoltageLevel().getNominalV();
+        loadTargetP += danglingLine.getP0();
+        loadTargetQ += danglingLine.getQ0();
     }
 
     @Override
@@ -33,13 +33,8 @@ public class LfDanglingLineBus extends AbstractFictitiousLfBus {
     }
 
     @Override
-    public double getLoadTargetP() {
-        return danglingLine.getP0() / PerUnit.SB;
-    }
-
-    @Override
-    public double getLoadTargetQ() {
-        return danglingLine.getQ0() / PerUnit.SB;
+    public boolean isFictitious() {
+        return true;
     }
 
     @Override
@@ -51,5 +46,7 @@ public class LfDanglingLineBus extends AbstractFictitiousLfBus {
     public void updateState(boolean reactiveLimits) {
         Networks.setPropertyV(danglingLine, v);
         Networks.setPropertyAngle(danglingLine, angle);
+
+        super.updateState(reactiveLimits);
     }
 }
