+++ src/main/java/com/powsybl/openloadflow/network/impl/LfStarBus.java
@@ -7,12 +7,11 @@
 package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.ThreeWindingsTransformer;
-import com.powsybl.openloadflow.network.AbstractFictitiousLfBus;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfStarBus extends AbstractFictitiousLfBus {
+public class LfStarBus extends AbstractLfBus {
 
     private final ThreeWindingsTransformer t3wt;
 
@@ -30,6 +29,11 @@ public class LfStarBus extends AbstractFictitiousLfBus {
     }
 
     @Override
+    public boolean isFictitious() {
+        return true;
+    }
+
+    @Override
     public double getNominalV() {
         return nominalV;
     }
@@ -38,5 +42,7 @@ public class LfStarBus extends AbstractFictitiousLfBus {
     public void updateState(boolean reactiveLimits) {
         Networks.setPropertyV(t3wt, v);
         Networks.setPropertyAngle(t3wt, angle);
+
+        super.updateState(reactiveLimits);
     }
 }
